refactor: consolidate native architecture
Collapse Rust public facades through api.rs, move PyO3 bindings under domain modules, and replace the old cleanup scratchpads with canonical architecture docs.

+++ crates/input/src/sample/alignment.rs
@@ -235,7 +235,7 @@ fn complete_multi_sample_array_indices(
         .collect()
 }
 
-fn sorted_sample_array_indices_by_sample_index(sample_indices: &[i64]) -> Vec<usize> {
+fn sorted_sample_array_indices_by_sample_index(sample_indices: &[usize]) -> Vec<usize> {
     let mut sorted_sample_array_indices: Vec<usize> = (0..sample_indices.len()).collect();
     sorted_sample_array_indices.sort_by_key(|sample_array_index| sample_indices[*sample_array_index]);
     sorted_sample_array_indices
